requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
openai==1.42.0
pdfminer.six==20231228
